refactor(codex_auth): move to llm/ module
Codex auth handles OpenAI Codex provider authentication — it's an LLM
communication layer concern. Primary caller is llm/openai.rs (4 refs).
No logic changes.
