+++ src/lib.rs
@@ -9,7 +9,6 @@ pub mod channels;
 pub mod config;
 pub mod error;
 pub mod llm;
-pub mod llm_profile;
 pub mod runtime;
 pub mod setup;
 pub mod skills;
