+++ src/tools/sanitizer.rs
@@ -3,7 +3,7 @@
 //! Config 由来のシークレット値と well-known パターンの二層リダクションにより、
 //! ツール出力に秘密情報が漏洩しないようマスクする。
 
-use crate::config::codex_auth::{is_codex_provider, resolve_codex_auth};
+use crate::llm::codex_auth::{is_codex_provider, resolve_codex_auth};
 use crate::config::Config;
 use crate::tools::ToolResult;
 
@@ -582,7 +582,7 @@ mod tests {
         let _guard = EnvVarGuard::set("HOME", dir.path())
             .also_set("OPENAI_CODEX_ACCESS_TOKEN", "")
             .also_set("CODEX_HOME", "");
-        crate::codex_auth::clear_auth_cache();
+        crate::llm::codex_auth::clear_auth_cache();
 
         let codex_dir = dir.path().join(".codex");
         std::fs::create_dir_all(&codex_dir).expect("create .codex dir");
